Refactor: Move ScanStateListItemBuilder to builder package
