+++ WordPress/src/test/java/org/wordpress/android/ui/jetpack/scan/ScanStateListItemBuilderTest.kt
@@ -5,6 +5,7 @@ import org.junit.Before
 import org.junit.Test
 import org.mockito.Mock
 import org.wordpress.android.BaseUnitTest
+import org.wordpress.android.ui.jetpack.scan.builders.ScanStateListItemBuilder
 import org.wordpress.android.ui.jetpack.scan.builders.ThreatItemBuilder
 import org.wordpress.android.ui.reader.utils.DateProvider
 import org.wordpress.android.ui.utils.HtmlMessageUtils
